issue-34: move PortType to separate file
