+++ wsdl-parser/src/parser/mod.rs
@@ -1,7 +1,7 @@
 use crate::parser::types::Definitions;
 
 mod constants;
-pub mod operation;
+pub mod port_type;
 pub mod types;
 
 pub fn parse(text: &str) {
